optimize pd dashboard writing
