+++ pkg/manager/member/tidbcluster_status_manager.go
@@ -14,20 +14,12 @@
 package member
 
 import (
-	"encoding/json"
 	"fmt"
 
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
-	"github.com/pingcap/tidb-operator/pkg/pdapi"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/klog"
-)
-
-const (
-	prometheusComponent = "prometheus"
-	grafanaComponent    = "grafana"
-	componentPrefix     = "/topology"
 )
 
 type TidbClusterStatusManager struct {
@@ -45,84 +37,7 @@ func (m *TidbClusterStatusManager) Sync(tc *v1alpha1.TidbCluster) error {
 }
 
 func (m *TidbClusterStatusManager) syncTidbMonitorRefAndKey(tc *v1alpha1.TidbCluster) error {
-	tm, err := m.syncTidbMonitorRef(tc)
-	if err != nil {
-		return err
-	}
-	err = m.syncDashboardMetricStorage(tc, tm)
-	if err != nil {
-		return err
-	}
 	return m.syncAutoScalerRef(tc)
-}
-
-func (m *TidbClusterStatusManager) syncTidbMonitorRef(tc *v1alpha1.TidbCluster) (*v1alpha1.TidbMonitor, error) {
-	if tc.Status.Monitor == nil {
-		return nil, nil
-	}
-	tmRef := tc.Status.Monitor
-	tm, err := m.deps.TiDBMonitorLister.TidbMonitors(tmRef.Namespace).Get(tmRef.Name)
-	if err != nil {
-		if errors.IsNotFound(err) {
-			tc.Status.Monitor = nil
-			err = nil
-		}
-		return nil, err
-	}
-	tcRef := tm.Spec.Clusters
-	if len(tcRef) < 1 {
-		tc.Status.Monitor = nil
-		return nil, nil
-	}
-	if len(tcRef[0].Namespace) < 1 {
-		tcRef[0].Namespace = tm.Namespace
-	}
-	if tcRef[0].Name != tc.Name || tcRef[0].Namespace != tc.Namespace {
-		tc.Status.Monitor = nil
-		return nil, nil
-	}
-	tc.Status.Monitor.GrafanaEnabled = true
-	if tm.Spec.Grafana == nil {
-		tc.Status.Monitor.GrafanaEnabled = false
-	}
-
-	return tm, nil
-}
-
-func (m *TidbClusterStatusManager) syncDashboardMetricStorage(tc *v1alpha1.TidbCluster, tm *v1alpha1.TidbMonitor) error {
-	if tc.Spec.PD == nil {
-		return nil
-	}
-	pdEtcdClient, err := m.deps.PDControl.GetPDEtcdClient(pdapi.Namespace(tc.Namespace), tc.Name, tc.IsTLSClusterEnabled())
-
-	if err != nil {
-		return err
-	}
-	if tc.IsTLSClusterEnabled() {
-		defer pdEtcdClient.Close()
-	}
-	var prometheusExist bool
-	var grafanaExist bool
-	if tc.Status.Monitor != nil {
-		prometheusExist = true
-		grafanaExist = tc.Status.Monitor.GrafanaEnabled
-	} else {
-		prometheusExist = false
-		grafanaExist = false
-	}
-
-	// sync prometheus key
-	err = syncComponent(prometheusExist, tm, prometheusComponent, 9090, pdEtcdClient)
-	if err != nil {
-		return err
-	}
-
-	// sync grafana key
-	err = syncComponent(grafanaExist, tm, grafanaComponent, 3000, pdEtcdClient)
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 func (m *TidbClusterStatusManager) syncAutoScalerRef(tc *v1alpha1.TidbCluster) error {
@@ -157,54 +72,6 @@ func (m *TidbClusterStatusManager) syncAutoScalerRef(tc *v1alpha1.TidbCluster) e
 		return nil
 	}
 	return nil
-}
-
-func syncComponent(exist bool, tm *v1alpha1.TidbMonitor, componentName string, port int, etcdClient pdapi.PDEtcdClient) error {
-	key := buildComponentKey(componentName)
-	if exist {
-		v, err := buildComponentValue(tm, componentName, port)
-		if err != nil {
-			klog.Error(err.Error())
-			return err
-		}
-		err = etcdClient.PutKey(key, v)
-		if err != nil {
-			klog.Error(err.Error())
-			return err
-		}
-	} else {
-		err := etcdClient.DeleteKey(key)
-		if err != nil {
-			klog.Error(err.Error())
-			return err
-		}
-	}
-	return nil
-}
-
-func buildComponentKey(component string) string {
-	return fmt.Sprintf("%s/%s", componentPrefix, component)
-}
-
-func buildComponentValue(tm *v1alpha1.TidbMonitor, componentName string, port int) (string, error) {
-	return buildEtcdValue(fmt.Sprintf("%s-%s.%s.svc", tm.Name, componentName, tm.Namespace), port)
-}
-
-type componentTopology struct {
-	IP   string `json:"ip"`
-	Port int    `json:"port"`
-}
-
-func buildEtcdValue(host string, port int) (string, error) {
-	topology := componentTopology{
-		IP:   host,
-		Port: port,
-	}
-	data, err := json.Marshal(topology)
-	if err != nil {
-		return "", err
-	}
-	return string(data), nil
 }
 
 type FakeTidbClusterStatusManager struct {
