+++ pkg/monitor/monitor/monitor_manager.go
@@ -19,6 +19,8 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/pingcap/tidb-operator/pkg/pdapi"
+
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	v1alpha1validation "github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1/validation"
 	"github.com/pingcap/tidb-operator/pkg/controller"
@@ -48,8 +50,11 @@ type MonitorManager struct {
 }
 
 const (
-	FailedSync  = "FailedSync"
-	SuccessSync = "SuccessSync"
+	FailedSync          = "FailedSync"
+	SuccessSync         = "SuccessSync"
+	prometheusComponent = "prometheus"
+	grafanaComponent    = "grafana"
+	componentPrefix     = "/topology"
 )
 
 func NewMonitorManager(deps *controller.Dependencies) *MonitorManager {
@@ -84,23 +89,10 @@ func (m *MonitorManager) SyncMonitor(monitor *v1alpha1.TidbMonitor) error {
 		if firstTc == nil && !tc.IsHeterogeneous() {
 			firstTc = tc
 		}
-		if tc.Status.Monitor != nil {
-			if tc.Status.Monitor.Name != monitor.Name || tc.Status.Monitor.Namespace != monitor.Namespace {
-				err := fmt.Errorf("tm[%s/%s]'s target tc[%s/%s] already referenced by TidbMonitor [%s/%s]", monitor.Namespace, monitor.Name, tc.Namespace, tc.Name, tc.Status.Monitor.Namespace, tc.Status.Monitor.Name)
-				m.deps.Recorder.Event(monitor, corev1.EventTypeWarning, FailedSync, err.Error())
-				return err
-			}
-		}
-		// TODO: Support validating webhook that forbids the tidbmonitor to update the monitorRef for the tidbcluster whose monitorRef has already
-		// been set by another TidbMonitor.
-		// Patch tidbcluster status first to avoid multi tidbmonitor monitoring the same tidbcluster
-		if !tc.IsHeterogeneous() {
-			if err := m.patchTidbClusterStatus(tc, monitor); err != nil {
-				message := fmt.Sprintf("Sync TidbMonitorRef into targetCluster[%s/%s] status failed, err:%v", tc.Namespace, tc.Name, err)
-				klog.Error(message)
-				m.deps.Recorder.Event(monitor, corev1.EventTypeWarning, FailedSync, err.Error())
-				return err
-			}
+		err = m.syncDashboardMetricStorage(tc, monitor)
+		if err != nil {
+			klog.Errorf("tidb[%s/%s] update dashboard pd monitor metrics error: %v", tc.Namespace, tc.Name, err)
+			continue
 		}
 	}
 
@@ -600,4 +592,82 @@ func (m *MonitorManager) patchPVClaimRef(pv *corev1.PersistentVolume, patchPvcNa
 		return err
 	}
 	return nil
+}
+
+func (m *MonitorManager) syncDashboardMetricStorage(tc *v1alpha1.TidbCluster, tm *v1alpha1.TidbMonitor) error {
+	if tc.Spec.PD == nil {
+		return nil
+	}
+	pdEtcdClient, err := m.deps.PDControl.GetPDEtcdClient(pdapi.Namespace(tc.Namespace), tc.Name, tc.IsTLSClusterEnabled())
+
+	if err != nil {
+		return err
+	}
+	if tc.IsTLSClusterEnabled() {
+		defer pdEtcdClient.Close()
+	}
+	var prometheusExist bool
+	var grafanaExist bool
+
+	grafanaExist = tm.Spec.Grafana != nil
+	// sync prometheus key
+	err = syncComponent(prometheusExist, tm, prometheusComponent, 9090, pdEtcdClient)
+	if err != nil {
+		return err
+	}
+
+	// sync grafana key
+	err = syncComponent(grafanaExist, tm, grafanaComponent, 3000, pdEtcdClient)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func syncComponent(exist bool, tm *v1alpha1.TidbMonitor, componentName string, port int, etcdClient pdapi.PDEtcdClient) error {
+	key := buildComponentKey(componentName)
+	if exist {
+		v, err := buildComponentValue(tm, componentName, port)
+		if err != nil {
+			klog.Error(err.Error())
+			return err
+		}
+		err = etcdClient.PutKey(key, v)
+		if err != nil {
+			klog.Error(err.Error())
+			return err
+		}
+	} else {
+		err := etcdClient.DeleteKey(key)
+		if err != nil {
+			klog.Error(err.Error())
+			return err
+		}
+	}
+	return nil
+}
+
+func buildComponentKey(component string) string {
+	return fmt.Sprintf("%s/%s", componentPrefix, component)
+}
+
+func buildComponentValue(tm *v1alpha1.TidbMonitor, componentName string, port int) (string, error) {
+	return buildEtcdValue(fmt.Sprintf("%s-%s.%s.svc", tm.Name, componentName, tm.Namespace), port)
+}
+
+type componentTopology struct {
+	IP   string `json:"ip"`
+	Port int    `json:"port"`
+}
+
+func buildEtcdValue(host string, port int) (string, error) {
+	topology := componentTopology{
+		IP:   host,
+		Port: port,
+	}
+	data, err := json.Marshal(topology)
+	if err != nil {
+		return "", err
+	}
+	return string(data), nil
 }
